# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "paths": {
        "/api/characters": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "List Characters",
                "description": "Get a paginated list of characters.",
                "operationId": "1797cf6529aea596ae858da8b98f8eff",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Character"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "No characters found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Characters"
                ],
                "summary": "Create Character",
                "description": "Create a new character.",
                "operationId": "9add7f48e7bda366ab508aeba92327a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Dead",
                                            "Alive",
                                            "Unknown"
                                        ]
                                    },
                                    "species": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "Female",
                                            "Male",
                                            "Genderless",
                                            "unknown"
                                        ]
                                    },
                                    "origin": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "link": {
                                                "type": "string",
                                                "format": "url"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "location": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "link": {
                                                "type": "string",
                                                "format": "url"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "episode": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "url"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Character created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Character"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/characters/{character}": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Show Character",
                "description": "Get details of a specific character.",
                "operationId": "a14d84704fee0e2def8b1c4cfc2d7e9a",
                "parameters": [
                    {
                        "name": "character",
                        "in": "path",
                        "description": "ID of the character",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Character"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Character not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Characters"
                ],
                "summary": "Update Character",
                "description": "Update an existing character.",
                "operationId": "a216ea8cf3f0798815490366b99cfe89",
                "parameters": [
                    {
                        "name": "character",
                        "in": "path",
                        "description": "Character ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Dead",
                                            "Alive",
                                            "Unknown"
                                        ]
                                    },
                                    "species": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "Female",
                                            "Male",
                                            "Genderless",
                                            "unknown"
                                        ]
                                    },
                                    "origin": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "link": {
                                                "type": "string",
                                                "format": "url"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "location": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "link": {
                                                "type": "string",
                                                "format": "url"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "episode": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "url"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Character updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Character"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Character not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Characters"
                ],
                "summary": "Delete Character",
                "description": "Delete a character.",
                "operationId": "aaaa0df57bb5ec0438234a433b25e8b8",
                "parameters": [
                    {
                        "name": "character",
                        "in": "path",
                        "description": "Character ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Character deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Authenticate the user and generate an access token.",
                "operationId": "5928fb15397c44cb50576a2dcd29cf7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Character": {
                "required": [
                    "id",
                    "name",
                    "status",
                    "species",
                    "type",
                    "gender",
                    "origin",
                    "location",
                    "image",
                    "episode",
                    "url",
                    "created"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "species": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "origin": {
                        "$ref": "#/components/schemas/Location"
                    },
                    "location": {
                        "$ref": "#/components/schemas/Location"
                    },
                    "image": {
                        "type": "string",
                        "format": "url"
                    },
                    "episode": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "url"
                        }
                    },
                    "url": {
                        "type": "string",
                        "format": "url"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Location": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "link": {
                        "type": "string",
                        "format": "url"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}